update Builder
The reason why the fetching result method can’t be declared inside the builder interface is that various builders may construct products that don’t have a common interface. Therefore, we don’t know what would be the return type for such a method. However, if we’re dealing with products from a single hierarchy, the fetching method can be safely added to the base interface.